fix: rename styles.ts to style.ts
